#686 Update project structure and docs for project
