+++ src/scribe_data/wiktionary/parse_translations.py
@@ -1305,7 +1305,7 @@ def _resolve_dump_path(
 
         import questionary
 
-        from scribe_data.cli.download import download_wiktionary_dumps
+        from scribe_data.cli.download.wiktionary_dump import download_wiktionary_dumps
 
         print(f"\nNo {wiktionary} dump found locally.")
         should_download = questionary.select(
