+++ tests/cli/contracts/test_cli_contracts_check.py
