+++ tests/cli/test_cli_get.py
