+++ tests/cli/test_cli_upgrade.py
