+++ tests/cli/test_cli_version.py
