--- tests/cli/test_interactive.py
@@ -1,260 +1,0 @@
-# SPDX-License-Identifier: GPL-3.0-or-later
-"""
-Tests for the CLI interactive mode functionality.
-"""
-
-import tempfile
-import unittest
-from pathlib import Path
-from unittest.mock import MagicMock, call, patch
-
-from prompt_toolkit.completion import WordCompleter
-
-from scribe_data.cli.interactive import (
-    ScribeDataConfig,
-    configure_settings,
-    display_summary,
-    prompt_for_data_types,
-    prompt_for_languages,
-    run_request,
-)
-
-
-class TestScribeDataInteractive(unittest.TestCase):
-    def setUp(self) -> None:
-        """
-        Set up test fixtures before each test method.
-        """
-        self.config = ScribeDataConfig()
-        # Mock the language_metadata and data_type_metadata.
-        self.config.languages = ["english", "spanish", "french"]
-        self.config.data_types = ["nouns", "verbs"]
-
-    def test_scribe_data_config_initialization(self) -> None:
-        """
-        Test ScribeDataConfig initialization.
-        """
-        self.assertEqual(self.config.selected_languages, [])
-        self.assertEqual(self.config.selected_data_types, [])
-        self.assertEqual(self.config.output_type, "json")
-        self.assertIsInstance(self.config.output_dir, Path)
-        self.assertFalse(self.config.overwrite)
-        self.assertFalse(self.config.configured)
-
-    @patch("scribe_data.cli.interactive.prompt")
-    @patch("scribe_data.cli.interactive.rprint")
-    def test_configure_settings_all_languages(
-        self, mock_rprint: MagicMock, mock_prompt: MagicMock
-    ) -> None:
-        """
-        Test configure_settings with 'All' languages selection.
-        """
-        # Set up mock responses.
-        responses = iter(
-            [
-                "All",  # languages
-                "nouns",  # data types
-                "json",  # output type
-                "",  # output directory (default)
-                "y",  # overwrite
-            ]
-        )
-        mock_prompt.side_effect = lambda *args, **kwargs: next(responses)
-
-        with patch("scribe_data.cli.interactive.config", self.config):
-            with patch("scribe_data.cli.interactive.display_summary"):
-                configure_settings()
-
-                self.assertEqual(self.config.selected_languages, self.config.languages)
-                self.assertEqual(self.config.selected_data_types, ["nouns"])
-                self.assertEqual(self.config.output_type, "json")
-                self.assertTrue(self.config.configured)
-
-    @patch("scribe_data.cli.interactive.prompt")
-    @patch("scribe_data.cli.interactive.rprint")
-    def test_configure_settings_specific_languages(
-        self, mock_rprint: MagicMock, mock_prompt: MagicMock
-    ) -> None:
-        """
-        Test configure_settings with specific language selection.
-        """
-        # Set up mock responses.
-        responses = iter(
-            [
-                "english, spanish",  # languages
-                "nouns, verbs",  # data types
-                "csv",  # output type
-                "/custom/path",  # output directory
-                "n",  # overwrite
-            ]
-        )
-        mock_prompt.side_effect = lambda *args, **kwargs: next(responses)
-
-        with patch("scribe_data.cli.interactive.config", self.config):
-            with patch("scribe_data.cli.interactive.display_summary"):
-                configure_settings()
-
-                self.assertEqual(self.config.selected_languages, ["english", "spanish"])
-                self.assertEqual(self.config.selected_data_types, ["nouns", "verbs"])
-                self.assertEqual(self.config.output_type, "csv")
-                self.assertEqual(self.config.output_dir.as_posix(), "/custom/path")
-                self.assertFalse(self.config.overwrite)
-
-    @patch("scribe_data.cli.interactive.get_data")
-    @patch("scribe_data.cli.interactive.tqdm")
-    @patch("scribe_data.cli.interactive.logger")
-    def test_run_request(
-        self, mock_logger: MagicMock, mock_tqdm: MagicMock, mock_get_data: MagicMock
-    ) -> None:
-        """
-        Test run_request functionality.
-        """
-        self.config.selected_languages = ["english"]
-        self.config.selected_data_types = ["nouns"]
-        self.config.configured = True
-
-        mock_get_data.return_value = True
-        mock_progress = MagicMock()
-        mock_tqdm.return_value.__enter__.return_value = mock_progress
-
-        with patch("scribe_data.cli.interactive.config", self.config):
-            run_request()
-
-            mock_get_data.assert_called_once_with(
-                languages=["english"],
-                data_types=["nouns"],
-                output_type=self.config.output_type,
-                output_dir=self.config.output_dir,
-                overwrite=self.config.overwrite,
-                interactive=True,
-            )
-
-    @patch("scribe_data.cli.interactive.prompt")
-    @patch("scribe_data.cli.interactive.rprint")
-    def test_request_total_lexeme(
-        self, mock_rprint: MagicMock, mock_prompt: MagicMock
-    ) -> None:
-        """
-        Test request_total_lexeme functionality.
-        """
-        # Set up mock responses.
-        mock_prompt.side_effect = [
-            "english, french",  # first call for languages
-            "nouns",  # first call for data types
-        ]
-
-        with patch("scribe_data.cli.interactive.config", self.config):
-            with patch(
-                "scribe_data.cli.interactive.list_all_languages",
-                return_value=["english", "french"],
-            ):
-                prompt_for_languages()
-                prompt_for_data_types()
-
-                # Verify the config was updated correctly.
-                self.assertEqual(self.config.selected_languages, ["english", "french"])
-                self.assertEqual(self.config.selected_data_types, ["nouns"])
-
-                # Verify prompt was called with correct arguments.
-                expected_calls = [
-                    call(
-                        "Select languages (comma-separated or 'All'): ",
-                        completer=unittest.mock.ANY,
-                        default="",
-                    ),
-                    call(
-                        "Select data types (comma-separated or 'All'): ",
-                        completer=unittest.mock.ANY,
-                        default="",
-                    ),
-                ]
-                mock_prompt.assert_has_calls(expected_calls, any_order=False)
-
-    @patch("rich.console.Console.print")
-    def test_display_summary(self, mock_print: MagicMock) -> None:
-        """
-        Test display_summary functionality.
-        """
-        self.config.selected_languages = ["english"]
-        self.config.selected_data_types = ["nouns"]
-        self.config.output_type = "json"
-
-        with patch("scribe_data.cli.interactive.config", self.config):
-            display_summary()
-            mock_print.assert_called()
-
-    def test_resolve_wiktionary_dump_path_from_subdirectory(self) -> None:
-        """
-        Find dumps when cwd is not the project root.
-        """
-        from scribe_data.cli.interactive import resolve_wiktionary_dump_path
-
-        with patch("os.getcwd") as mock_getcwd:
-            with tempfile.TemporaryDirectory() as tmp:
-                root = Path(tmp)
-                dump_dir = root / "scribe_data_wiktionary_dumps_export"
-                json_dir = root / "scribe_data_json_export"
-                dump_dir.mkdir()
-                json_dir.mkdir()
-                dump_file = dump_dir / "dewiktionary-pages-articles.xml.bz2"
-                dump_file.write_bytes(b"x")
-
-                mock_getcwd.return_value = str(json_dir)
-                resolved = resolve_wiktionary_dump_path(
-                    "german",
-                    "scribe_data_wiktionary_dumps_export",
-                )
-
-                self.assertEqual(resolved, dump_file.resolve())
-
-    def test_create_word_completer(self) -> None:
-        """
-        Test create_word_completer functionality.
-        """
-        from scribe_data.cli.interactive import create_word_completer
-
-        # Test without 'All' option.
-        options = ["english", "spanish", "french"]
-        completer = create_word_completer(options, include_all=False)
-        self.assertIsInstance(completer, WordCompleter)
-        self.assertEqual(completer.words, options)
-
-        # Test with 'All' option.
-        completer_with_all = create_word_completer(options, include_all=True)
-        self.assertEqual(completer_with_all.words, ["All"] + options)
-
-    @patch(
-        "scribe_data.cli.interactive.resolve_wiktionary_dump_path",
-        return_value=Path("/dump/path"),
-    )
-    @patch("scribe_data.wiktionary.parse_translations.parse_wiktionary_translations")
-    @patch("scribe_data.cli.interactive.prompt")
-    @patch("scribe_data.cli.interactive.prompt_for_languages")
-    @patch("scribe_data.cli.interactive.questionary.select")
-    def test_start_interactive_mode_translations(
-        self,
-        mock_select,
-        mock_prompt_languages,
-        mock_prompt,
-        mock_parse_wiktionary,
-        mock_resolve_dump,
-    ):
-        from scribe_data.cli.interactive import config, start_interactive_mode
-
-        mock_select.return_value.ask.side_effect = ["translations"]
-        mock_prompt.side_effect = [
-            "german",
-            "/dump/path",
-            "scribe_data_wiktionary_json_export",
-            "false",
-        ]
-        config.selected_languages = ["english"]
-
-        start_interactive_mode(operation="translations")
-
-        mock_parse_wiktionary.assert_called_once_with(
-            target_languages=["english"],
-            wiktionary_dump_path=Path("/dump/path"),
-            output_dir=Path("scribe_data_wiktionary_json_export"),
-            overwrite=False,
-        )
