+++ tests/unicode/test_unicode_generate_emoji_keywords.py
